+++ packages/wonder-blocks-core/src/hooks/use-is-mounted.ts
@@ -1,22 +1,28 @@
 import * as React from "react";
 
-import {useOnMountEffect} from "./use-on-mount-effect";
+import {useIsomorphicLayoutEffect} from "./use-isomorphic-layout-effect";
 
 /**
  * Hook to provide a function for determining component mounted state.
  *
- * NOTE: Based on https://github.com/juliencrn/usehooks-ts/blob/d5f3de88cc319c790f2a4e90ba6a8904298957a5/src/useIsMounted/useIsMounted.ts
+ * NOTE: Inspired by https://github.com/juliencrn/usehooks-ts/blob/d5f3de88cc319c790f2a4e90ba6a8904298957a5/src/useIsMounted/useIsMounted.ts
+ *
+ * This returns a callback to access the mounted state as components should not
+ * be reactive to the mounted state, but instead should be accessing it in
+ * callbacks to guard against setting state on an unmounted component.
  *
  * @returns {() => boolean} A function that returns the component mounted state.
  */
 export const useIsMounted = (): (() => boolean) => {
-    const isMounted = React.useRef<boolean>(false);
-    useOnMountEffect(() => {
-        isMounted.current = true;
-        return () => {
-            isMounted.current = false;
-        };
-    });
+    const isMountedRef = React.useRef<boolean>(false);
+    const isMounted = React.useCallback(() => isMountedRef.current, []);
 
-    return React.useCallback(() => isMounted.current, []);
+    useIsomorphicLayoutEffect(() => {
+        isMountedRef.current = true;
+        return () => {
+            isMountedRef.current = false;
+        };
+    }, []);
+
+    return isMounted;
 };