+++ packages/wonder-blocks-labeled-field/src/components/labeled-field.tsx
@@ -1,0 +1,4 @@
+// Something so that the file is not empty
+const LabeledField = "LabeledField";
+
+export default LabeledField;