+++ packages/wonder-blocks-theming/src/hooks/use-scoped-theme.ts
@@ -1,5 +1,21 @@
 import * as React from "react";
+import {ThemeSwitcherContext} from "../utils/theme-switcher-context";
 
-export default function useScopedTheme<T>(themeContext: React.Context<T>): T {
-    return React.useContext(themeContext);
+type ScopedTheme<T> = {
+    /**
+     * The theme object.
+     */
+    theme: T;
+    /**
+     * The theme name.
+     */
+    themeName: string;
+};
+
+export default function useScopedTheme<T>(
+    themeContext: React.Context<T>,
+): ScopedTheme<T> {
+    const theme = React.useContext(themeContext);
+    const themeName = React.useContext(ThemeSwitcherContext) ?? "default";
+    return {theme, themeName};
 }