+++ packages/wonder-blocks-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @khanacademy/wonder-blocks-core
 
+## 6.0.2
+
+### Patch Changes
+
+-   f19da46e: Allow 'title' on View and typography components
+
+## 6.0.1
+
+### Patch Changes
+
+-   1920feb8: `RenderStateRoot` now wraps children in a React fragment
+
 ## 6.0.0
 
 ### Major Changes