+++ packages/wonder-blocks-grid/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @khanacademy/wonder-blocks-grid
 
+## 2.0.19
+
+### Patch Changes
+
+-   Updated dependencies [f19da46e]
+    -   @khanacademy/wonder-blocks-core@6.0.2
+    -   @khanacademy/wonder-blocks-layout@2.0.19
+
+## 2.0.18
+
+### Patch Changes
+
+-   Updated dependencies [1920feb8]
+    -   @khanacademy/wonder-blocks-core@6.0.1
+    -   @khanacademy/wonder-blocks-layout@2.0.18
+
 ## 2.0.17
 
 ### Patch Changes