+++ packages/wonder-blocks-layout/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-layout
 
+## 2.0.19
+
+### Patch Changes
+
+-   Updated dependencies [f19da46e]
+    -   @khanacademy/wonder-blocks-core@6.0.2
+
+## 2.0.18
+
+### Patch Changes
+
+-   Updated dependencies [1920feb8]
+    -   @khanacademy/wonder-blocks-core@6.0.1
+
 ## 2.0.17
 
 ### Patch Changes