+++ packages/wonder-blocks-link/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @khanacademy/wonder-blocks-link
 
+## 5.0.3
+
+### Patch Changes
+
+-   Updated dependencies [f19da46e]
+    -   @khanacademy/wonder-blocks-core@6.0.2
+    -   @khanacademy/wonder-blocks-clickable@4.0.3
+    -   @khanacademy/wonder-blocks-icon@2.1.3
+
+## 5.0.2
+
+### Patch Changes
+
+-   Updated dependencies [8dc4a5a3]
+-   Updated dependencies [1920feb8]
+    -   @khanacademy/wonder-blocks-clickable@4.0.2
+    -   @khanacademy/wonder-blocks-core@6.0.1
+    -   @khanacademy/wonder-blocks-icon@2.1.2
+
 ## 5.0.1
 
 ### Patch Changes